Set minimum Perl version in metadata
Since this addition requires at least version 6.48 of
ExtUtils::MakeMaker, the version check in Makefile.PL has been bumped to
at least this version.

Even though the minimum version of Perl that the module code supports is
5.6.0, the version is set to 5.8.0 here since the test code requires at
least this version (since it uses the `use utf8` pragma).  Since the
version metadata information is used in things such as CPAN Testers it
will be helpful there, however by *not* setting the minimum version
explicitly in the code means that someone wanting to install and run
this module on a 5.6.x Perl would still be able to do so.